refactor: remove unused mock files and update test setup
- Deleted mock files: mocks.ts, app-environment.ts, and setup.ts.
- Moved app-environment.ts to tests/mocks/ and updated references.
- Updated test setup to use new mock structure and paths.
- Refactored integration tests to utilize new mock functions and structures.
- Enhanced test coverage for admin API, authentication, profile management, and song choice functionalities.
