+++ app/package.json
@@ -33,6 +33,8 @@
 		"@sveltejs/vite-plugin-svelte": "^6.0.0",
 		"@tailwindcss/typography": "^0.5.15",
 		"@tailwindcss/vite": "^4.0.0",
+		"@testing-library/jest-dom": "^6.8.0",
+		"@testing-library/svelte": "^5.2.8",
 		"@types/node": "^24.3.0",
 		"eslint": "^9.18.0",
 		"eslint-config-prettier": "^10.0.1",
